updated userID ig dimenssions

+++ api_ig_get/src/main/java/com/dolphpire/api/action/igaccount/posts/PostsIGAccountAction.java
@@ -9,7 +9,7 @@ public class PostsIGAccountAction
         this.mPostsIGAccount = new PostsIGAccount();
     }
 
-    public PostsIGAccountAction withUserID(int igUserID) {
+    public PostsIGAccountAction withUserID(long igUserID) {
         this.mPostsIGAccount.setUserID(String.valueOf(igUserID));
         return this;
     }
